Refactor mobile footer styles and add skip links
Adjust the styles for the mobile footer to include margin, padding, and a flex layout. Also, add skip links to allow users to navigate directly to the main content and footer sections.

Refactor main.css: Update styles for pointer, icons, and cart number

Update the styles in main.css to improve the appearance of pointer, icons, and cart number elements.

Refactor orders.html and detailArticle.html: Add skip links and update footer

Add skip links to the orders.html and detailArticle.html pages to allow users to navigate directly to the main content and footer sections. Also, update the footer with copyright information and social media links.

Refactor article.ts: Update property name from "catégorie" to "categorie"

Update the property name from "catégorie" to "categorie" in the article.ts file for consistency and clarity.

Refactor main.css: Adjust padding for user icon in navigation

Adjust the padding for the user icon in the navigation section of the main.css file to improve the layout and alignment.